requirements.txt
black==25.1.0 ; python_version >= "3.11" and python_full_version < "3.12.0"
certifi==2025.7.9 ; python_version >= "3.11" and python_full_version < "3.12.0"
charset-normalizer==3.4.2 ; python_version >= "3.11" and python_full_version < "3.12.0"
click==8.2.1 ; python_version >= "3.11" and python_full_version < "3.12.0"
colorama==0.4.6 ; python_version >= "3.11" and python_full_version < "3.12.0" and (platform_system == "Windows" or sys_platform == "win32")
contourpy==1.3.2 ; python_version >= "3.11" and python_full_version < "3.12.0"
cycler==0.12.1 ; python_version >= "3.11" and python_full_version < "3.12.0"
defusedxml==0.7.1 ; python_version >= "3.11" and python_full_version < "3.12.0"
filelock==3.18.0 ; python_version >= "3.11" and python_full_version < "3.12.0"
flake8==7.3.0 ; python_version >= "3.11" and python_full_version < "3.12.0"
fonttools==4.58.5 ; python_version >= "3.11" and python_full_version < "3.12.0"
fsspec==2025.5.1 ; python_version >= "3.11" and python_full_version < "3.12.0"
ghp-import==2.1.0 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
idna==3.10 ; python_version >= "3.11" and python_full_version < "3.12.0"
iniconfig==2.1.0 ; python_version >= "3.11" and python_full_version < "3.12.0"
isort==6.0.1 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
jinja2==3.1.6 ; python_version >= "3.11" and python_full_version < "3.12.0"
kiwisolver==1.4.8 ; python_version >= "3.11" and python_full_version < "3.12.0"
markdown==3.8.2 ; python_version >= "3.11" and python_full_version < "3.12.0"
markupsafe==3.0.2 ; python_version >= "3.11" and python_full_version < "3.12.0"
matplotlib==3.10.3 ; python_version >= "3.11" and python_full_version < "3.12.0"
mccabe==0.7.0 ; python_version >= "3.11" and python_full_version < "3.12.0"
mergedeep==1.3.4 ; python_version >= "3.11" and python_full_version < "3.12.0"
mkdocs-get-deps==0.2.0 ; python_version >= "3.11" and python_full_version < "3.12.0"
mkdocs==1.6.1 ; python_version >= "3.11" and python_full_version < "3.12.0"
mpmath==1.3.0 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
mypy-extensions==1.1.0 ; python_version >= "3.11" and python_full_version < "3.12.0"
networkx==3.5 ; python_version >= "3.11" and python_full_version < "3.12.0"
numpy==2.2.6 ; python_version >= "3.11" and python_full_version < "3.12.0"
nvidia-cublas-cu12==12.6.4.1 ; python_version >= "3.11" and python_full_version < "3.12.0" and platform_system == "Linux" and platform_machine == "x86_64"
nvidia-cuda-cupti-cu12==12.6.80 ; python_version >= "3.11" and python_full_version < "3.12.0" and platform_system == "Linux" and platform_machine == "x86_64"
nvidia-cuda-nvrtc-cu12==12.6.77 ; python_version >= "3.11" and python_full_version < "3.12.0" and platform_system == "Linux" and platform_machine == "x86_64"
nvidia-cuda-runtime-cu12==12.6.77 ; python_version >= "3.11" and python_full_version < "3.12.0" and platform_system == "Linux" and platform_machine == "x86_64"
nvidia-cudnn-cu12==9.5.1.17 ; python_version >= "3.11" and python_full_version < "3.12.0" and platform_system == "Linux" and platform_machine == "x86_64"
nvidia-cufft-cu12==11.3.0.4 ; python_version >= "3.11" and python_full_version < "3.12.0" and platform_system == "Linux" and platform_machine == "x86_64"
nvidia-cufile-cu12==1.11.1.6 ; python_version >= "3.11" and python_full_version < "3.12.0" and platform_system == "Linux" and platform_machine == "x86_64"
nvidia-curand-cu12==10.3.7.77 ; python_version >= "3.11" and python_full_version < "3.12.0" and platform_system == "Linux" and platform_machine == "x86_64"
nvidia-cusolver-cu12==11.7.1.2 ; python_version >= "3.11" and python_full_version < "3.12.0" and platform_system == "Linux" and platform_machine == "x86_64"
nvidia-cusparse-cu12==12.5.4.2 ; python_version >= "3.11" and python_full_version < "3.12.0" and platform_system == "Linux" and platform_machine == "x86_64"
nvidia-cusparselt-cu12==0.6.3 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0" and platform_system == "Linux" and platform_machine == "x86_64"
nvidia-nccl-cu12==2.26.2 ; python_version >= "3.11" and python_full_version < "3.12.0" and platform_system == "Linux" and platform_machine == "x86_64"
nvidia-nvjitlink-cu12==12.6.85 ; python_version >= "3.11" and python_full_version < "3.12.0" and platform_system == "Linux" and platform_machine == "x86_64"
nvidia-nvtx-cu12==12.6.77 ; python_version >= "3.11" and python_full_version < "3.12.0" and platform_system == "Linux" and platform_machine == "x86_64"
opencv-python==4.12.0.88 ; python_version >= "3.11" and python_full_version < "3.12.0"
packaging==25.0 ; python_version >= "3.11" and python_full_version < "3.12.0"
pandas==2.3.1 ; python_version >= "3.11" and python_full_version < "3.12.0"
pathspec==0.12.1 ; python_version >= "3.11" and python_full_version < "3.12.0"
pillow==11.3.0 ; python_version >= "3.11" and python_full_version < "3.12.0"
pip==25.1.1 ; python_version >= "3.11" and python_full_version < "3.12.0"
platformdirs==4.3.8 ; python_version >= "3.11" and python_full_version < "3.12.0"
pluggy==1.6.0 ; python_version >= "3.11" and python_full_version < "3.12.0"
psutil==7.0.0 ; python_version >= "3.11" and python_full_version < "3.12.0"
py-cpuinfo==9.0.0 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
pycodestyle==2.14.0 ; python_version >= "3.11" and python_full_version < "3.12.0"
pyflakes==3.4.0 ; python_version >= "3.11" and python_full_version < "3.12.0"
pygments==2.19.2 ; python_version >= "3.11" and python_full_version < "3.12.0"
pyparsing==3.2.3 ; python_version >= "3.11" and python_full_version < "3.12.0"
pytest==8.4.1 ; python_version >= "3.11" and python_full_version < "3.12.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.11" and python_full_version < "3.12.0"
python-dotenv==1.1.1 ; python_version >= "3.11" and python_full_version < "3.12.0"
pytz==2025.2 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
pyyaml-env-tag==1.1 ; python_version >= "3.11" and python_full_version < "3.12.0"
pyyaml==6.0.2 ; python_version >= "3.11" and python_full_version < "3.12.0"
requests==2.32.4 ; python_version >= "3.11" and python_full_version < "3.12.0"
scipy==1.16.0 ; python_version >= "3.11" and python_full_version < "3.12.0"
setuptools==80.9.0 ; python_version >= "3.11" and python_full_version < "3.12.0" and platform_system == "Linux" and platform_machine == "x86_64"
six==1.17.0 ; python_version >= "3.11" and python_full_version < "3.12.0"
supervision==0.25.1 ; python_version == "3.11"
sympy==1.14.0 ; python_version >= "3.11" and python_full_version < "3.12.0"
torch==2.7.1 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0"
torchvision==0.22.1 ; python_version >= "3.11" and python_full_version < "3.12.0"
tqdm==4.67.1 ; python_version >= "3.11" and python_full_version < "3.12.0"
triton==3.3.1 ; python_full_version >= "3.11.0" and python_full_version < "3.12.0" and platform_system == "Linux" and platform_machine == "x86_64"
typing-extensions==4.14.1 ; python_version >= "3.11" and python_full_version < "3.12.0"
tzdata==2025.2 ; python_version >= "3.11" and python_full_version < "3.12.0"
ultralytics-thop==2.0.14 ; python_version >= "3.11" and python_full_version < "3.12.0"
ultralytics==8.3.163 ; python_version >= "3.11" and python_full_version < "3.12.0"
urllib3==2.5.0 ; python_version >= "3.11" and python_full_version < "3.12.0"
watchdog==6.0.0 ; python_version >= "3.11" and python_full_version < "3.12.0"
